+++ src/core/attrsgenerator.cpp
@@ -18,9 +18,6 @@
  *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include <functional>
-#include <memory>
-
 #include <QFile>
 #include <QFileInfo>
 #include <QTextStream>