+++ src/test/tst_prg.cpp
@@ -33,11 +33,11 @@ private slots:
     void initTestCase() {}
     void cleanupTestCase() {}
     void tst_prg();
-    void tst_randB();
-    void tst_randD();
-    void tst_randI();
-    void tst_randS();
-    void tst_randF();
+    void tst_bernoulli();
+    void tst_uniformDouble();
+    void tst_uniformInt();
+    void tst_uniformSizeT();
+    void tst_uniformFloat();
 };
 
 void TestPRG::tst_prg()
@@ -45,37 +45,37 @@ void TestPRG::tst_prg()
     PRG* prg1 = new PRG(921);
     PRG* prg2 = new PRG(921);
 
-    QCOMPARE(prg1->randD(), prg2->randD(0.0, 1.0));
-    QCOMPARE(prg1->randD(), prg2->randD(1.0));
-    QCOMPARE(prg1->randI(123), prg2->randI(0, 123));
+    QCOMPARE(prg1->uniform(), prg2->uniform(0.0, 1.0));
+    QCOMPARE(prg1->uniform(), prg2->uniform(1.0));
+    QCOMPARE(prg1->uniform(123), prg2->uniform(0, 123));
 
-    double d = prg1->randD();
-    int i = prg1->randI(123);
-    QCOMPARE(d, prg2->randD());
-    QCOMPARE(i, prg2->randI(123));
-    QVERIFY(!qFuzzyCompare(d, prg2->randD()));
-    QVERIFY(i != prg2->randI(123));
+    double d = prg1->uniform();
+    int i = prg1->uniform(123);
+    QCOMPARE(d, prg2->uniform());
+    QCOMPARE(i, prg2->uniform(123));
+    QVERIFY(!qFuzzyCompare(d, prg2->uniform()));
+    QVERIFY(i != prg2->uniform(123));
 
     delete prg1;
     prg1 = new PRG(922);
-    QVERIFY(!qFuzzyCompare(d, prg1->randD()));
-    QVERIFY(i != prg1->randI(123));
+    QVERIFY(!qFuzzyCompare(d, prg1->uniform()));
+    QVERIFY(i != prg1->uniform(123));
 
     delete prg1;
     delete prg2;
 
     prg1 = new PRG(0);
-    for (int i = 0; i < 1000; ++i) prg1->randD();
+    for (int i = 0; i < 1000; ++i) prg1->uniform();
     prg2 = new PRG(0);
-    for (int i = 0; i < 1000; ++i) prg2->randD();
-    QCOMPARE(prg1->randD(), prg2->randD());
-    QCOMPARE(prg1->randI(10), prg2->randI(10));
+    for (int i = 0; i < 1000; ++i) prg2->uniform();
+    QCOMPARE(prg1->uniform(), prg2->uniform());
+    QCOMPARE(prg1->uniform(10), prg2->uniform(10));
 
     delete prg1;
     delete prg2;
 }
 
-void TestPRG::tst_randB()
+void TestPRG::tst_bernoulli()
 {
     auto prg = std::unique_ptr<PRG>(new PRG(0));
     const int size = 1000;
@@ -84,8 +84,8 @@ void TestPRG::tst_randB()
     int trues1 = 0;
     int trues2 = 0;
     for (size_t i = 0; i < size; ++i) {
-        if (prg->randB()) ++trues1;
-        if (prg->randBernoulli()) ++trues2;
+        if (prg->bernoulli()) ++trues1;
+        if (prg->bernoulli(.5)) ++trues2;
     }
     QVERIFY(trues1 > 400 && trues1 < 600);
     QVERIFY(trues2 > 400 && trues2 < 600);
@@ -93,22 +93,18 @@ void TestPRG::tst_randB()
     // general cases
     auto tst = [size, &prg](double p) {
         int trues1 = 0;
-        int trues2 = 0;
         for (size_t i = 0; i < size; ++i) {
-            if (prg->randBernoulli(p)) ++trues1;
-            if (prg->randB(p)) ++trues2;
+            if (prg->bernoulli(p)) ++trues1;
         }
         QVERIFY(trues1 > size*(p-0.1)); // 10% margin
         QVERIFY(trues1 < size*(p+0.1));
-        QVERIFY(trues2 > size*(p-0.1));
-        QVERIFY(trues2 < size*(p+0.1));
     };
     for (double p = 0.0; p <= 1.0; p+=0.1) {
         tst(p);
     }
 }
 
-void TestPRG::tst_randD()
+void TestPRG::tst_uniformDouble()
 {
     PRG* prg = new PRG(0);
     double v, min, max;
@@ -116,27 +112,27 @@ void TestPRG::tst_randD()
     // [0, 1)
     min = 0.0;
     max = 1.0;
-    v = prg->randD(min, max);
+    v = prg->uniform(min, max);
     QVERIFY(v >= min && v < max);
-    v = prg->randD();
+    v = prg->uniform();
     QVERIFY(v >= min && v < max);
 
     // negative
     min = -100.0;
     max = -50.0;
-    v = prg->randD(min, max);
+    v = prg->uniform(min, max);
     QVERIFY(v >= min && v < max);
 
     // same number
     min = 123.45;
     max = min;
-    v = prg->randD(min, max);
+    v = prg->uniform(min, max);
     QVERIFY(v == min);
 
     delete prg;
 }
 
-void TestPRG::tst_randI()
+void TestPRG::tst_uniformInt()
 {
     PRG* prg = new PRG(0);
     int v, min, max;
@@ -144,27 +140,27 @@ void TestPRG::tst_randI()
     // [0, 10)
     min = 0;
     max = 10;
-    v = prg->randI(min, max);
+    v = prg->uniform(min, max);
     QVERIFY(v >= min && v <= max);
-    v = prg->randI(max);
+    v = prg->uniform(max);
     QVERIFY(v >= min && v <= max);
 
     // negative
     min = -100;
     max = -50;
-    v = prg->randI(min, max);
+    v = prg->uniform(min, max);
     QVERIFY(v >= min && v <= max);
 
     // same number
     min = 123;
     max = min;
-    v = prg->randI(min, max);
+    v = prg->uniform(min, max);
     QVERIFY(v == min);
 
     delete prg;
 }
 
-void TestPRG::tst_randS()
+void TestPRG::tst_uniformSizeT()
 {
     PRG* prg = new PRG(0);
     size_t v, min, max;
@@ -172,31 +168,29 @@ void TestPRG::tst_randS()
     // [0, 10)
     min = 0;
     max = 10;
-    v = prg->randS(min, max);
+    v = prg->uniform(min, max);
     QVERIFY(v >= min && v <= max);
-
-    // negative
-    min = -100;
-    max = -50;
-    v = prg->randS(min, max);
+    v = prg->uniform(max);
     QVERIFY(v >= min && v <= max);
 
     // same number
     min = 123;
     max = min;
-    v = prg->randS(min, max);
+    v = prg->uniform(min, max);
     QVERIFY(v == min);
 
     // large numbers
     min = UINT64_MAX - 10;
     max = UINT64_MAX;
-    v = prg->randS(min, max);
+    v = prg->uniform(min, max);
     QVERIFY(v >= min && v <= max);
+    v = prg->uniform(max);
+    QVERIFY(v >= 0 && v <= max);
 
     delete prg;
 }
 
-void TestPRG::tst_randF()
+void TestPRG::tst_uniformFloat()
 {
     PRG* prg = new PRG(0);
     float v, min, max;
@@ -204,24 +198,21 @@ void TestPRG::tst_randF()
     // [0, 10)
     min = 0;
     max = 10;
-    v = prg->randF(min, max);
+    v = prg->uniform(min, max);
     QVERIFY(v >= min && v <= max);
-    v = prg->randF(max);
+    v = prg->uniform(max);
     QVERIFY(v >= min && v <= max);
-    v = prg->randF();
-    max = 1;
-    QVERIFY(v >= min && v < max);
 
     // negative
     min = -100;
     max = -50;
-    v = prg->randF(min, max);
+    v = prg->uniform(min, max);
     QVERIFY(v >= min && v <= max);
 
     // same number
     min = 123;
     max = min;
-    v = prg->randF(min, max);
+    v = prg->uniform(min, max);
     QVERIFY(v == min);
 
     delete prg;